+++ src/cow-react/api/gnosisProtocol/errors/OperatorError.ts
