+++ src/custom/state/gnosisSafe/updater.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react'
 import { useIsGnosisSafeWallet } from 'hooks/useWalletInfo'
-import { getSafeInfo } from 'api/gnosisSafe'
+import { getSafeInfo } from 'cow-react/api/gnosisSafe'
 import { useWeb3React } from '@web3-react/core'
 import { gnosisSafeAtom } from 'state/gnosisSafe/atoms'
 import { useUpdateAtom } from 'jotai/utils'
